rename menu mod to system_tray
